refactor: [피드백반영] Results객체 파싱을 외부에서 동작 후 생성
