+++ src/main/java/nextstep/ladder/domain/Results.java
@@ -5,37 +5,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Results {
-    public static final String RESULT_EMPTY_MSG = "결과는 공백 일 수 없습니다.";
-    public static final String RESULT_DELIMITER = ",";
-    public static final String INVALID_RESULT_COUNT_MSG = "참가자의 수와 결과의 수는 다를 수 없습니다.";
     private final String[] results;
 
-    public Results(final int countOfPerson, final String resultText) {
-        emptyCheck(resultText);
-
-        final String[] results = resultText.split(RESULT_DELIMITER);
-
-        resultNumberCheck(countOfPerson, results);
-
+    public Results(String[] results) {
         this.results = results;
     }
-
-    private static void resultNumberCheck(final int countOfPerson, final String[] results) {
-        if (results.length != countOfPerson) {
-            throw new IllegalArgumentException(INVALID_RESULT_COUNT_MSG);
-        }
-    }
-
-    private void emptyCheck(final String resultText) {
-        if (isNullOrEmpty(resultText)) {
-            throw new IllegalArgumentException(RESULT_EMPTY_MSG);
-        }
-    }
-
-    private boolean isNullOrEmpty(final String resultText) {
-        return resultText == null || resultText.isEmpty();
-    }
-
 
     public String get(final Position endPosition) {
         return this.results[endPosition.get()];
