--- src/test/java/nextstep/ladder/domain/LineTest.java
@@ -1,11 +1,0 @@
-package nextstep.ladder.domain;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class LineTest {
-
-}
